SDC-561. Refactor BaseKafkaSource to remove self-parent relationship
Reviewed-on: https://review.streamsets.net/344
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ container/src/main/java/com/streamsets/pipeline/util/SystemProcessImpl.java
@@ -107,8 +107,10 @@ public class SystemProcessImpl implements SystemProcess {
     if (errorTailer != null) {
       errorTailer.close();
     }
-    error.delete();
-    output.delete();
+    if (!Boolean.getBoolean("sdc.testing-mode")) {
+      error.delete();
+      output.delete();
+    }
     kill(5000);
   }
   @Override
